requirements.txt
streamlit
openai
PyPDF2
python_docx
langchain_community
langchain_huggingface
sentence_transformers

